+++ remote_print/src/shared.rs
@@ -40,7 +40,7 @@ pub enum PrintInfoStatus {
     Stopping,
     Stopped,
     Complete,
-    CleckingFile,
+    CheckingFile,
     FinalRetract,
     Canceled, // maybe?
     Unknown(u8),
@@ -62,7 +62,7 @@ impl<'de> Deserialize<'de> for PrintInfoStatus {
             7 => Self::Stopping,
             8 | 14 => Self::Stopped, // 14 happens when the stop command is sent
             9 | 16 => Self::Complete,
-            10 => Self::CleckingFile,
+            10 => Self::CheckingFile,
             12 => Self::FinalRetract,
             13 => Self::Canceled,
             other => Self::Unknown(other),