+++ Changelog.md
@@ -22,6 +22,7 @@
 - Only create mesh acceleration structures once per mesh, not per instance
 - Fix rendering of 'Send to Printer' menu
 - Add 'Help' menu to toolbar (has links to website, source code, and bug tracker)
+- Don't trigger model shortcuts when typing in a text field
 
 ## v0.9.1 &mdash; August 28th, 2026